MDEV-33533 Crash at execution of DELETE when trying to use rowid filter
This bug could cause a crash of the server when executing multi-delete if
the optimization using rowid filters was allowed. The same problem could
be observed for a delete if it was converted into multi-delete, in
particular for a delete containing IN subqueries. The problem could appear
only if the query used InnoDB tables.

Multi-delete as well as multu-update usually is executed in two phases:
first a corresponding select is performed to find the records to be
affected by the command, then delete/update of these records is carried
out. If select uses a rowid filter to access the table whose records are
to be deleted/updated then the filter should be disabled for the second
phase. Otherwise it can confuse the used storage engine what could be seen
at execution of the query reported for the bug.

Disabling of the rowid filter at the second phase of multi-update execution
was added earlier, but execution of multi-delete missed it.

Approved by Oleksandr Byelkin <sanja@mariadb.com>